successfully updating database now

+++ app/src/main/java/com/example/pricify/model/Wishlist.kt
@@ -4,5 +4,5 @@ data class Wishlist (
     val name: String,
     val items: List<Item>,
     val item_count: Int = items.size,
-    val total_price: Int = items.sumOf { it.price },
+    val total_price: Double = items.sumOf { it.price },
 )
